remove advice to reorder experiences
There was no consensus about whether or not you should reorder your most recent experiences to highlight relevant experience. Some say this is good, others say it's confusing because employers expect experience to be in reverse chronological order. Since this checklist is supposed to be contain "no debate" criteria, I'll remove it.